[ffx][config]Making "file" processing return 1 string
This changes the `--process file` handling in `ffx config get`
to return the first path configured that exists. Before this
change, if the config value was an array, it would return a
single value if only 1 of the entries in the list existed, but
if more than one existed, it would return an array, making it
very difficult to properly handle.

This change makes it always return a single entry.

Bug: b/341777920
Change-Id: I415d1f091547fca4fdac41a6e7ae3f0c2d058f04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051241
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>