[starnix][selinux] Replace has_permission with has_permissions
Replace the generic `IntoIterator` argument to `has_permissions()`
with a slice argument, to simplify the implementation and allow
Rust to take advantage of the fact that sets of permissions will
always be expressed as static arrays.

`has_permission()` is removed, since `has_permissions()` is
expected to be sufficiently simple to be fast-enough for single-
permission checks.

Bug: 335397745
Change-Id: I4baf1062e10109bb074dcaf451425f9490dfa813
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028973
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>