[ramdisk] remove GetName support
For some reason, ramdisks supported the Partition.GetName fidl call,
returning their name as "ramdisk-x". This makes sense in a vacuum, but
since ramdisks are almost always used in place of real disks, they
should act like real disks wherever possible, and real disks don't
support this call, it is generally reserved for block devices served out
of partition maps.

This is required for a change to fshost which checks the result of
GetName when matching a device for fvm or fxblob. The change succeeds on
real devices but the integration tests failed because ramdisks didn't
follow the same pattern.

Change-Id: I6faa4e2ec0e3e47079a816131b1962e17ff8eafc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048392
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>