[inspect][graph] Improve initial metadata event recording
Instead of iterating twice over all the metadata, we can construct the
event node as we record the metadata and then reparent it when creating
the event. This removes the need of requiring that the iterator is
clonable as well and simplifies the implementation and performance of
the follow-up change which brings nested metadata support.

Bug: 340342858
Change-Id: I6462c2dccf4b30502ae799e05574d6560f622487
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050233
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>