[tee] Add manager component
This adds a TA Manager component which is responsible for instantiating
TAs according to the configured policy and monitoring them. Once we add
the ability to route capabilities to specific TAs the manager will be
responsible for this as well. The first version of the TA contains
configuration parsing logic to determine which TAs exist and what
implementation we want to use for each UUID.

This iteration has support for knowing if a TA supports multiple
instances but it doesn't use this information yet. As such, it does not
correctly handle multiple connections to the same UUID.

Bug: b/332956721

Change-Id: Ic7300bb2e350662646c6bdd1cf65a518c0e3dde4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042480
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>