[platform-device] Update platform device naming convention
Use the name of the platform device as the name of the driver framework
device created by that platform device.

Create an allowlist for platform devices that should use the legacy name
format when naming their child device. End goal is to remove this
allowlist.

Bug: 340283894
Change-Id: Ic25c1ff114fae7521e755d7ad3fa03933c5534b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046618
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>