[netstack3] Add NatContext
Define an execution context for performing NAT, which currently includes
one method that provides the ability to get the best local address for
communicating with a particular remote, over a particular interface.
This is useful currently for Redirect NAT in picking a destination
address for an incoming packet, but it will also be useful in the future
for Masquerade NAT (a special case of SNAT), where it will conversely be
used to select an appropriate source address for an outgoing packet.

Bug: 333419001
Change-Id: I2f4e1df4e712f36e10047a8486df88827ce4cd1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049953
Reviewed-by: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>