[msd-arm-mali] Prefix some GPU logs with type
Generally, one log line at the beginning of a dump is important to help
identify the exact type of error that caused the dump. Prefix that line
in a standard way to make it easier to identify.

Change-Id: Ie3583d00b2d1ca0963f81f1449949d9ab2d49d0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052552
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>