[netstack3] Change local delivery on loopback
... allowing delivery when the destination address is contained within a
subnet assigned to the loopback device (rather than an exact match with
the assigned address).

This conforms to behavior of Netstack2 and Linux.

While here, add a helper to `TestApi` to add the loopback device, and
migrate existing integration tests to use it.

Fixed: 341767462
Change-Id: I497ce471a2fa9dd72746360c21ca3ccdabe0dfe9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051332
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>