[kernel][chainlock] Be sure to watch the proper CPU when backing off.
When encountering a ChainLock conflict and backing off, remember which
CPU we had the conflict on before unregistering our active CLT and
waiting for the lock we wanted to be released.  Failure to do this can
lead to a situation where we migrate to a different CPU which _just
happens_ to have the same conflict ID as the CPU we just came from.
If we stay on this CPU forever, and nothing ever generates a new
conflict on our new CPU, we will just stay stuck forever.

See the detailed write-up in b/341370301 if you are interested in the
precise sequence.

Fixed: 341370301
Change-Id: Ic65fb99f8027ae54be3777206363d9add7f0f622
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052096
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>