[vim3-devicetree] Add missing NNA metadata
External sram base metadata was missing from nna node. This is not the
right way to pass this data to underlying drivers. For now, a temporary
visitor is added to pass this data from the board driver. Eventually
this data should either be passed directly to the driver consuming it
from pbus or should be populated by the aml-nna driver.

Bug: 342174810
Change-Id: Id1642c0a5188695403e2a04f219043974147aaa2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052220
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>