[fxfs] Improve SimpleLayerWriter code.
I discovered by chance that this code was doing deallocations as part of
writing because it was using skip() (which does a flush) to jump over a
small number of zero bytes. I've rewritten the code to fill a buffer and
write in one go. Seems a bit faster now and enabled me to write a
disk-full test to boot!. :P

Change-Id: I2c8ab377f24364306b3ff7453b3eaf3c48ff54da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051714
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Chris Suter <csuter@google.com>