[component_manager] Moniker-based storage paths use fixed id 0
Instead of using a variable id for dynamic instances in a storage path,
use 0 as the fixed instance id. The reasons for using nonzero instance
ids are historical.

For collections that use `persistent_storage: true`, this is a no-op
change: these paths already zero their instance ids. (This is why we set
the instance id to 0 instead of removing it altogether, to retain
compatibility for these paths.)

For collections with `persistent_storage: false`, the paths will change,
but there should be no compatibility break because all storage
capabilities routed to such collections are backed by in memory
filesystems.

This change is a precursor to removing instanced monikers altogether.

Bug: 297390990
Change-Id: Id908778be18e9327aea47a61ce0f6c101618e28f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051355
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Gary Bressler <geb@google.com>