[netstack3] Apply mirror trait pattern to datagram
UDP and ICMP socket's blanket implementations of the datagram socket
contexts do not work in a split-crate world.

Apply the mirror trait pattern on the datagram spec type (which is
always local) so we can resolve that.

To slightly reduce the boilerplate, a type alias is extracted for bound
sockets, and the Converter associated types are removed in favor of an
impl return from the trait plus a marker trait to tie things together.

The mirror traits are all in their own file so the datagram module
doesn't need to consider their existence.

The rename on `converter` to `ds_converter` and `nds_converter` was
necessary to help the type inference, which got broken in a number of
places with the introduction of new traits. I find that in the end this
also helps the reader know whether this is a dual stack context or not.

Finally, a sad side effect of the mirror traits is that all the tests in
the datagram module need turbofish specifiers to tell which datagram
spec is being used. We're swallowing the ugliness for now, that should
come undone if we extract all the standalone functions from datagram
into an Api struct like the other modules have, which I plan on doing
post-split.

Bug: 338448926
Change-Id: I60ac0fe8d6dada658c8e67916ff65ddf617192ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053795
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>