[component_manager] program input dictionary additions is not optional
Removes the `Optional` wrapping program input dictionary additions,
because an empty dictionary is functionally identical to a `None` value.
This allows us to pass the dictionary of additions directly into the
dictionary construction code, where it will retain a reference to the
additions dictionary and thus not need to upgrade a component reference
when handling requests that may involve it.

Change-Id: Iafaef15003d752e74466f7d39bbff1f9837e36dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044835
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>