[display][coordinator] Move dispatcher ownership to CoordinatorDriver.
Previously, in the display coordinator driver, the Controller class
creates a fdf Dispatcher as the main event loop. This causes some
different behaviors on Controller destruction: in DFv2 drivers, the
dispatcher is destroyed by the driver framework before Stop(); while
in tests, the dispatcher is destroyed by the Controller's Stop() method.

This change fixes this issue by moving dispatcher ownership to
CoordinatorDriver, which is the class that interacts with the driver
framework. Controller now only holds an unowned reference to the
dispatcher, and it requires the dispatcher be completely stopped
when Stop() is called (which is guaranteed by DFv2 driver lifecycle).
Tests now need to manually stop the dispatcher and wait until the
dispatcher is fully stopped before calling Controller::Stop().

Test: no-op on core.nelson and core.vim3.
      "display-tool squares" worked.
Test: display-coordinator-unittests
Bug: 341156315
Change-Id: I4a4d3289bb10b9cb28a515847edce698629504d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1055692
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>