[netstack3] Implement Redirect NAT
Implement the entry point for invoking NAT from IP filtering hooks,
which attempts to configure NAT for a connection if it has not yet been
configured, by running the packet through the installed NAT routines,
and then if NAT has been configured, performs NAT on the packet.

Currently the only supported type of NAT is Redirect NAT, a special case
of DNAT. DNAT is configured for a connection by rewriting the source
fields of the reply tuple of a conntrack connection, which serves as a
guide to rewrite future packets seen in both directions on the
connection.

If a destination port range is specified in a Redirect action, the
netstack attempts to find a port in the range that would not cause the
rewritten reply tuple to conflict with another tuple in the conntrack
table. If it fails to find one, the packet is dropped.

Bug: 333419001
Test: netstack3-filter-test
Change-Id: If83a17983e1771bba6d53dbfbefd37e73b81468a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050372
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>