[aml-uart] Add lease management based on timer
Introduce a timer in the driver so that the driver can hold the wake
lease for a certain time period before dropping it. The driver resets
the timer each time when the wake vector interrupt is asserted and
drops the lease then the timer times out.

Bug: b/339233136
Test: New unittests and observation on begonia(The driver doesn't drop
the lease while packets coming in, and drops the lease a while after
the last packet).

Multiply: aml-uart-dfv2-test

Change-Id: Icc047c9609d39f9973b52385752e3e6cce0abe05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049252
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>