[kernel][sched] Make multiple attempts to find a CPU for a thread.
Change the semantics of the "active" bit in the mp maintained active
CPU mask.  The mask may still be observed without holding lock, but to
mutate the bit corresponding to CPU/Scheduler, the Schedueler's "queue
lock" must be held.

The implication of this is that, while _not_ holding the lock, the
values in the mask amount to a "best guess", useful in
`Scheduler::FindTargetCpu` when searching for a CPU to run a given
thread on.  That said, when it finally comes to actually queue'ing a
thread onto a target CPU, we must first hold that CPU's Scheduler's
lock, and re-verify that the CPU is actually active, potentially
backing off and re-running `FindTargetCpu` in the case that our CPU
was de-activated after we selected it, but before we actually queued
our thread on it.

Fixed: 342649393
Change-Id: Id00a9b7084109d2111f2e6655199921a1e1ddcc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056093
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>