[netstack3] Remove box iterator from hot path
Extract with_assigned_ipv4_addr_subnets to a test utility so we don't
rely on the box in the hot path for IPv4 address selection. It's still
used in a single test, but keeping it around for symmetry with the IPv6
version which is more widely used and has more intricate logic.

Change-Id: I3730b91dbe79a29d16e28c335cb27a870034b322
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054710
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>