[vim3-devicetree] Suppress dtc warnings related to display address
There are intentionally two display nodes hdmi and dsi which both use
the same registers. This is because we do not know at build time which
of the two displays would be connected. Ideally, it would be nice to
have both the detection and driver logic in a single driver which would
avoid the duplicate register usage warning. But there is no plan to
change these drivers and hence suppressing dtc the warning.

Change-Id: I11ee09a8c0392ae8df834220b7586f49d7931007
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057899
Reviewed-by: Braden Kell <bradenkell@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>