[kernel] Handle memory levels in MemoryWatchdog
Moves all of the memory level handling and tracking from the PmmNode to
the MemoryWatchdog. The different memory levels, except for the OOM
state, had no meaning for the PMM and were just used to generate
signals. This was complexity and logic that did not need to be in the
PMM and is only leveraged by the MemoryWatchdog.

Having the PMM notifications be more dynamic, instead of the static
reclamation levels previously, allows additional flexibility in the
future for synthesizing more complicated pressure levels based on a
mixture of PMM free and other available memory sources.

Run-All-Tests: True

Change-Id: I3d16cdc2b0f43ab6c54db0ee809838c886ae7607
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031612
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>