[fdf] Implement bootup tracker
Add a bootup tracker in the driver manager to keep
track of when all driver bind and start activity
are completed, indicating bootup is done.

This CL adds a BootupWatcher service which the tracker
implements to notify when bootup is done.

The tracker also logs an warning if there are any
hanging binding and start requests.

Bug: 333714068
Tests: bootup-tracker-test
Change-Id: I30426536d6389aa6a59fab55318c6b897a2e103b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038333
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>