[perfcompare] Fix "--expected_sample_size" check for non-perfcompare builds
Fix the recently-added "--expected_sample_size" check so that it gets
applied in the case where only one results directory is passed to
perfcompare.py, which is the common case for when results are
displayed in non-perfcompare builds.

The early return in ComparePerf() stopped this check from being
applied in that case.  I accidentally broke the check for that case by
moving it to the end of the ComparePerf() function.

Bug: 343997506
Run-All-Tests: True
Change-Id: I344ccd7afe61aff857edeb2b110f29968e968f58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059276
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>