[sdk][build] Apply sdk_sub_build_verbose to checked subprocesses
Previously, the gn's output in sub-builds was being captured and only
shown if the sub-build failed. This meant that print statements would
not be displayed, which could be confusing.

To address this, do not capture the output when --verbose is passed to
subbuild.py as is the case when the GN arg `sdk_sub_build_verbose` is
true.

Bug: 337028633
Test: Observe "IDK_SUBBUILD_api_20_x64 [0/1](1) Regenerating ninja files" when running when `fx build final_fuchsia_idk` while `sdk_sub_build_verbose=true`.
Change-Id: I429a4075d59f35ca67ab13faf7c400d0a6aa0254
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058718
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>