[fxfs] Change the way we replay the journal
This CL changes how we replay to be more intuitive/simpler at the cost
of some potential performance (during replay).

Prior to this change, we would replay the mutations first and then open
the immuatble layers.  This changes things so that we open first and
then replay the mutations.  Doing it this way means that as we replay,
we don't need to keep track of deltas to apply when we open; we can just
operate the same as we do during live operation.

The downside to this is that we have to perform extra passes over the
transactions that we need to replay.  We need two more passes: one to
apply all the root parent store's mutations, and then another one to
apply the root store's mutations.  After that, we can apply all the
other mutations.

Bug: b/341928701
Change-Id: Ibf8ee7c14d13249bbcffdab9b6ae9f893c025d1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058913
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>